Move tests to ci
