--- tests/ci_tests/fast_test.py
@@ -1,1 +1,0 @@
-print('The mindsdb server has not continous integration testing, we should change this in the future !')
